New page
Got tired of the todoList so did this as a side.

Saw someone do this with Yu-Gi-Oh cards an though this might actually be fun. And it was.

Got a little confused about the difference between append and appendChild. Still don't understand the difference between just doing: element.href = something. And element.setAttribute('href', 'something'). Or are they the same? Also tried element.insertAdjacentElement() but it just seemed unnecessary for my use of it as it was much more specific.

Added:
Pokedex.html
pokedex-app.js